requirements.txt
opencv-python==4.5.4.60
keras=2.7.0
pillow=9.2.0